--- a/chia/full_node/mempool.py
+++ b/chia/full_node/mempool.py
@@ -1,18 +1,26 @@
 from __future__ import annotations
 
+import sqlite3
+from dataclasses import dataclass
 from datetime import datetime
 from enum import Enum
-from typing import Dict, List, Optional
+from typing import Dict, Iterator, List, Optional
 
-from sortedcontainers import SortedDict
-
+from chia.consensus.cost_calculator import NPCResult
 from chia.full_node.fee_estimation import FeeMempoolInfo, MempoolInfo, MempoolItemInfo
 from chia.full_node.fee_estimator_interface import FeeEstimatorInterface
-from chia.types.blockchain_format.coin import Coin
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.clvm_cost import CLVMCost
 from chia.types.mempool_item import MempoolItem
-from chia.util.ints import uint64
+from chia.types.spend_bundle import SpendBundle
+from chia.util.chunks import chunks
+from chia.util.db_wrapper import SQLITE_MAX_VARIABLE_NUMBER
+from chia.util.ints import uint32, uint64
+
+# We impose a limit on the fee a single transaction can pay in order to have the
+# sum of all fees in the mempool be less than 2^63. That's the limit of sqlite's
+# integers, which we rely on for computing fee per cost as well as the fee sum
+MEMPOOL_ITEM_FEE_LIMIT = 2**50
 
 
 class MempoolRemoveReason(Enum):
@@ -21,48 +29,117 @@ class MempoolRemoveReason(Enum):
     POOL_FULL = 3
 
 
+@dataclass(frozen=True)
+class InternalMempoolItem:
+    spend_bundle: SpendBundle
+    npc_result: NPCResult
+    height_added_to_mempool: uint32
+
+
 class Mempool:
+    _db_conn: sqlite3.Connection
+    # it's expensive to serialize and deserialize G2Element, so we keep those in
+    # this separate dictionary
+    _items: Dict[bytes32, InternalMempoolItem]
+
     def __init__(self, mempool_info: MempoolInfo, fee_estimator: FeeEstimatorInterface):
-        self._spends: Dict[bytes32, MempoolItem] = {}
-        self._sorted_spends: SortedDict = SortedDict()
+        self._db_conn = sqlite3.connect(":memory:")
+        self._items = {}
+
+        with self._db_conn:
+            # name means SpendBundle hash
+            # assert_height may be NIL
+            self._db_conn.execute(
+                """CREATE TABLE tx(
+                name BLOB PRIMARY KEY,
+                cost INT NOT NULL,
+                fee INT NOT NULL,
+                assert_height INT,
+                fee_per_cost REAL GENERATED ALWAYS AS (fee / cost) VIRTUAL)
+                """
+            )
+            self._db_conn.execute("CREATE INDEX fee_sum ON tx(fee)")
+            self._db_conn.execute("CREATE INDEX cost_sum ON tx(cost)")
+            self._db_conn.execute("CREATE INDEX feerate ON tx(fee_per_cost)")
+
+            # This table maps coin IDs to spend bundles hashes
+            self._db_conn.execute(
+                """CREATE TABLE spends(
+                coin_id BLOB NOT NULL,
+                tx BLOB NOT NULL,
+                UNIQUE(coin_id, tx))
+                """
+            )
+            self._db_conn.execute("CREATE INDEX spend_by_coin ON spends(coin_id)")
+            self._db_conn.execute("CREATE INDEX spend_by_bundle ON spends(tx)")
+
         self.mempool_info: MempoolInfo = mempool_info
         self.fee_estimator: FeeEstimatorInterface = fee_estimator
-        self._removal_coin_id_to_spendbundle_ids: Dict[bytes32, List[bytes32]] = {}
-        self._total_mempool_cost: CLVMCost = CLVMCost(uint64(0))
-        self._total_mempool_fees: int = 0
+
+    def __del__(self) -> None:
+        self._db_conn.close()
+
+    def _row_to_item(self, row: sqlite3.Row) -> MempoolItem:
+        name = bytes32(row[0])
+        fee = int(row[1])
+        assert_height = row[2]
+        item = self._items[name]
+
+        return MempoolItem(
+            item.spend_bundle, uint64(fee), item.npc_result, name, uint32(item.height_added_to_mempool), assert_height
+        )
 
     def total_mempool_fees(self) -> int:
-        return self._total_mempool_fees
+        with self._db_conn:
+            cursor = self._db_conn.execute("SELECT SUM(fee) FROM tx")
+            val = cursor.fetchone()[0]
+            return uint64(0) if val is None else uint64(val)
 
     def total_mempool_cost(self) -> CLVMCost:
-        return self._total_mempool_cost
+        with self._db_conn:
+            cursor = self._db_conn.execute("SELECT SUM(cost) FROM tx")
+            val = cursor.fetchone()[0]
+            return CLVMCost(uint64(0) if val is None else uint64(val))
 
-    def all_spends(self) -> List[MempoolItem]:
-        return list(self._spends.values())
+    def all_spends(self) -> Iterator[MempoolItem]:
+        with self._db_conn:
+            cursor = self._db_conn.execute("SELECT name, fee, assert_height FROM tx")
+            for row in cursor:
+                yield self._row_to_item(row)
 
     def all_spend_ids(self) -> List[bytes32]:
-        return list(self._spends.keys())
+        with self._db_conn:
+            cursor = self._db_conn.execute("SELECT name FROM tx")
+            return [bytes32(row[0]) for row in cursor]
 
-    def spends_by_feerate(self) -> List[MempoolItem]:
-        ret: List[MempoolItem] = []
-        for spends_with_fpc in reversed(self._sorted_spends.values()):
-            ret.extend(spends_with_fpc.values())
-        return ret
+    # TODO: move "process_mempool_items()" into this class in order to do this a
+    # bit more efficiently
+    def spends_by_feerate(self) -> Iterator[MempoolItem]:
+        with self._db_conn:
+            cursor = self._db_conn.execute("SELECT name, fee, assert_height FROM tx ORDER BY fee_per_cost DESC")
+            for row in cursor:
+                yield self._row_to_item(row)
 
     def size(self) -> int:
-        return len(self._spends)
+        with self._db_conn:
+            cursor = self._db_conn.execute("SELECT Count(name) FROM tx")
+            val = cursor.fetchone()
+            return 0 if val is None else int(val[0])
 
     def get_spend_by_id(self, spend_bundle_id: bytes32) -> Optional[MempoolItem]:
-        return self._spends.get(spend_bundle_id, None)
+        with self._db_conn:
+            cursor = self._db_conn.execute("SELECT name, fee, assert_height FROM tx WHERE name=?", (spend_bundle_id,))
+            row = cursor.fetchone()
+            return None if row is None else self._row_to_item(row)
 
+    # TODO: we need a bulk lookup function like this too
     def get_spends_by_coin_id(self, spent_coin_id: bytes32) -> List[MempoolItem]:
-        spend_bundle_ids = self._removal_coin_id_to_spendbundle_ids.get(spent_coin_id)
-        if spend_bundle_ids is None:
-            return []
-        ret: List[MempoolItem] = []
-        for spend_bundle_id in spend_bundle_ids:
-            ret.append(self._spends[spend_bundle_id])
-        return ret
+        with self._db_conn:
+            cursor = self._db_conn.execute(
+                "SELECT name, fee, assert_height FROM tx WHERE name in (SELECT tx FROM spends WHERE coin_id=?)",
+                (spent_coin_id,),
+            )
+            return [self._row_to_item(row) for row in cursor]
 
     def get_min_fee_rate(self, cost: int) -> float:
         """
@@ -70,16 +147,21 @@ class Mempool:
         """
 
         if self.at_full_capacity(cost):
-            current_cost = self._total_mempool_cost
+            # TODO: make MempoolItem.cost be CLVMCost
+            current_cost = int(self.total_mempool_cost())
 
             # Iterates through all spends in increasing fee per cost
-            fee_per_cost: float
-            for fee_per_cost, spends_with_fpc in self._sorted_spends.items():
-                for spend_name, item in spends_with_fpc.items():
-                    current_cost -= item.cost
+            with self._db_conn:
+                cursor = self._db_conn.execute("SELECT cost,fee_per_cost FROM tx ORDER BY fee_per_cost ASC")
+
+                item_cost: int
+                fee_per_cost: float
+                for item_cost, fee_per_cost in cursor:
+                    current_cost -= item_cost
                     # Removing one at a time, until our transaction of size cost fits
                     if current_cost + cost <= self.mempool_info.max_size_in_cost:
                         return fee_per_cost
+
             raise ValueError(
                 f"Transaction with cost {cost} does not fit in mempool of max cost {self.mempool_info.max_size_in_cost}"
             )
@@ -90,61 +172,67 @@ class Mempool:
         """
         Removes an item from the mempool.
         """
-        for spend_bundle_id in items:
-            item: Optional[MempoolItem] = self._spends.get(spend_bundle_id)
-            if item is None:
-                continue
-            assert item.name == spend_bundle_id
-            removals: List[Coin] = item.removals
-            for rem in removals:
-                rem_name: bytes32 = rem.name()
-                self._removal_coin_id_to_spendbundle_ids[rem_name].remove(spend_bundle_id)
-                if len(self._removal_coin_id_to_spendbundle_ids[rem_name]) == 0:
-                    del self._removal_coin_id_to_spendbundle_ids[rem_name]
-            del self._spends[item.name]
-            del self._sorted_spends[item.fee_per_cost][item.name]
-            dic = self._sorted_spends[item.fee_per_cost]
-            if len(dic.values()) == 0:
-                del self._sorted_spends[item.fee_per_cost]
-            self._total_mempool_cost = CLVMCost(uint64(self._total_mempool_cost - item.cost))
-            self._total_mempool_fees = self._total_mempool_fees - item.fee
-            assert self._total_mempool_cost >= 0
-            info = FeeMempoolInfo(self.mempool_info, self._total_mempool_cost, self._total_mempool_fees, datetime.now())
-            if reason != MempoolRemoveReason.BLOCK_INCLUSION:
-                self.fee_estimator.remove_mempool_item(
-                    info, MempoolItemInfo(item.cost, item.fee, item.height_added_to_mempool)
-                )
+        removed_items: List[MempoolItemInfo] = []
+        if reason != MempoolRemoveReason.BLOCK_INCLUSION:
+
+            for spend_bundle_ids in chunks(items, SQLITE_MAX_VARIABLE_NUMBER):
+                args = ",".join(["?"] * len(spend_bundle_ids))
+                with self._db_conn:
+                    cursor = self._db_conn.execute(
+                        f"SELECT name, cost, fee FROM tx WHERE name in ({args})", spend_bundle_ids
+                    )
+                    for row in cursor:
+                        name = bytes32(row[0])
+                        internal_item = self._items[name]
+                        item = MempoolItemInfo(int(row[1]), int(row[2]), internal_item.height_added_to_mempool)
+                        removed_items.append(item)
+
+        for name in items:
+            self._items.pop(name)
+
+        for spend_bundle_ids in chunks(items, SQLITE_MAX_VARIABLE_NUMBER):
+            args = ",".join(["?"] * len(spend_bundle_ids))
+            with self._db_conn:
+                self._db_conn.execute(f"DELETE FROM tx WHERE name in ({args})", spend_bundle_ids)
+                self._db_conn.execute(f"DELETE FROM spends WHERE tx in ({args})", spend_bundle_ids)
+
+        if reason != MempoolRemoveReason.BLOCK_INCLUSION:
+            info = FeeMempoolInfo(
+                self.mempool_info, self.total_mempool_cost(), self.total_mempool_fees(), datetime.now()
+            )
+            for iteminfo in removed_items:
+                self.fee_estimator.remove_mempool_item(info, iteminfo)
 
     def add_to_pool(self, item: MempoolItem) -> None:
         """
         Adds an item to the mempool by kicking out transactions (if it doesn't fit), in order of increasing fee per cost
         """
 
+        assert item.fee < MEMPOOL_ITEM_FEE_LIMIT
         assert item.npc_result.conds is not None
 
-        while self.at_full_capacity(item.cost):
-            # Val is Dict[hash, MempoolItem]
-            fee_per_cost, val = self._sorted_spends.peekitem(index=0)
-            to_remove: MempoolItem = list(val.values())[0]
-            self.remove_from_pool([to_remove.name], MempoolRemoveReason.POOL_FULL)
+        # TODO: this block could be simplified by removing all items in a single
+        # SQL query. Or at least figure out which items to remove and then
+        # remove them all in a single call to remove_from_pool()
+        with self._db_conn:
+            while self.at_full_capacity(item.cost):
+                # pick the item with the lowest fee per cost to remove
+                cursor = self._db_conn.execute("SELECT name FROM tx ORDER BY fee_per_cost ASC LIMIT 1")
+                name = bytes32(cursor.fetchone()[0])
+                self.remove_from_pool([name], MempoolRemoveReason.POOL_FULL)
 
-        self._spends[item.name] = item
+            self._db_conn.execute(
+                "INSERT INTO tx VALUES(?, ?, ?, ?)", (item.name, item.cost, item.fee, item.assert_height)
+            )
 
-        # _sorted_spends is Dict[float, Dict[bytes32, MempoolItem]]
-        if item.fee_per_cost not in self._sorted_spends:
-            self._sorted_spends[item.fee_per_cost] = {}
+            all_coin_spends = [(s.coin_id, item.name) for s in item.npc_result.conds.spends]
+            self._db_conn.executemany("INSERT INTO spends VALUES(?, ?)", all_coin_spends)
 
-        self._sorted_spends[item.fee_per_cost][item.name] = item
+            self._items[item.name] = InternalMempoolItem(
+                item.spend_bundle, item.npc_result, item.height_added_to_mempool
+            )
 
-        for coin in item.removals:
-            coin_id = coin.name()
-            if coin_id not in self._removal_coin_id_to_spendbundle_ids:
-                self._removal_coin_id_to_spendbundle_ids[coin_id] = []
-            self._removal_coin_id_to_spendbundle_ids[coin_id].append(item.name)
-
-        self._total_mempool_cost = CLVMCost(uint64(self._total_mempool_cost + item.cost))
-        self._total_mempool_fees = self._total_mempool_fees + item.fee
-        info = FeeMempoolInfo(self.mempool_info, self._total_mempool_cost, self._total_mempool_fees, datetime.now())
+        info = FeeMempoolInfo(self.mempool_info, self.total_mempool_cost(), self.total_mempool_fees(), datetime.now())
         self.fee_estimator.add_mempool_item(info, MempoolItemInfo(item.cost, item.fee, item.height_added_to_mempool))
 
     def at_full_capacity(self, cost: int) -> bool:
@@ -152,4 +240,4 @@ class Mempool:
         Checks whether the mempool is at full capacity and cannot accept a transaction with size cost.
         """
 
-        return self._total_mempool_cost + cost > self.mempool_info.max_size_in_cost
+        return self.total_mempool_cost() + cost > self.mempool_info.max_size_in_cost