# overview

The objective of this change is to allow to *reliably* add more sophisticated rules to the mempool.

This patch adds an sqlite database as the *index* for mempool items. In order to not repeat values from the index in the items themselves, it uses an `InternalMempoolItem` with fewer fields. the `_items` dictionary (which stores the mempool items) is kept in sync with the sqlite database, referencing items in the dict, by *name* (SpendBundle hash).

# rationale

The immediate features to add are the `assert_before_height` and `assert_before_seconds` fields on mempool items, ensuring they expire once those heights or timestamps have been reached.

I'm also envisioning a future where the "pending_cache" and "conflict_cache" are also folded into the database. That would also simplify the mempool data structure (or really, offload the complex parts to sqlite).

With that, I also imagine it would be easy to add support for `ASSERT_SECONDS`, which currently is not supported by the mempool (presumably to keep complexity in check).

# benefits

Fundamentally, the mempool is a multi-index container. A collection of items that we want efficient access to via different fields and ordering. e.g. we want to be able to iterate by fee rate, look up by spend bundle hash, look up by spent coin ID.

There is no native support for multi index cointainers in python, and implementing one is complex and error prone. Using a sqlite database greatly simplifies the data structure we need to support, by offloading all indices. This is especially significant as we add more indices (as mentioned under rationale). The complexity of the mempool will scale better with the number of indices and logic we add to the mempool.

# drawbacks

The main drawback of using a sqlite database is a slight performance degradation. The least common operation (farming a block) is slowed down and takes about 30% more time. See benchmarks below.

Another noteworthy consequence of this patch is that to total fee (i.e. the sum of all fees of all spend bundles) is not allowed to exceed `INT64_MAX`. This is because sqlite uses signed int64 as its internal representation for integers, so in order for the `fee_per_cost` field to be computed (and implied), sqlite need to be able to interpret the integers. This also allows the total fee and total cost to simply be sqlite queries for `SUM(<field>)`.

This patch, additionally, adds a fee limit of 2^50 mojos for an individual `MempoolItem`.

I don't think this restriction is problematic, since you would need astronomical fees in order to hit this ceiling, and it would only push out other transactions for one block (and then require more astronomical fees to keep it up).

# benchmarks

## Ubuntu AMD64 (CPU)

| function | before | after | after / before
| --- | --- | --- | --- |
| add_spend_bundle() | 6.3161s | 6.4140s | 101.55% |
| add_spend_bundle() replace | 6.2743s | 6.3782s | 101.66 % |
| create_bundle_from_mempool() | 4.5330s | 5.9528s | 131.32 % |
| new_peak() optimized | 16.6200s | 17.2157s | 103.58 % |
| new_peak() reorg | 16.2587s | 16.9127s | 104 % |

## Ubuntu AMD64 (memory)

The RAM pressure is pretty crude, just measuring the heap pressure of the mempool benchmark.
The benchmark logic itself probably contributes half of the memory pressure.

| metric | before | after | after / before |
| --- | --- | --- |--- |
| Max RSS | 78864 | 79908 | 101.32% |

## Raspberry PI (CPU)

| function | before | after | after / before
| --- | --- | --- | --- |
| add_spend_bundle() | 18.99s | 19.09s | 100.5 % |
| add_spend_bundle() replace | 18.99s | 19.09s | 100.54 % |
| create_bundle_from_mempool() | 16.98s | 21.53s | 126.78 % |
| new_peak() optimized | 47.12 s | 46.48s | 98.63 % |
| new_peak() reorg | 45.69 s | 46.42s | 101.6 % |

## Ubuntu amd64 (CPU)

current `main`:
```
Profiling add_spend_bundle()
  output written to: mempool-add-mt.png
  time: 6.3161s
  per call: 3.16ms

Profiling add_spend_bundle() with replace-by-fee
  output written to: mempool-replace-mt.png
  time: 6.2743s
  per call: 3.14ms

Profiling create_bundle_from_mempool()
  output written to: mempool-create-mt.png
  time: 4.5330s
  per call: 9.07ms

Profiling new_peak() (optimized)
  output written to: mempool-new-peak-mt.png
  time: 16.6200s
  per call: 8.31ms

Profiling new_peak() (reorg)
  output written to: mempool-new-peak-reorg-mt.png
  time: 16.2587s
  per call: 8.13ms
```

with sqlite:
```
Profiling add_spend_bundle()
  output written to: mempool-add-mt.png
  time: 6.4140s
  per call: 3.21ms

Profiling add_spend_bundle() with replace-by-fee
  output written to: mempool-replace-mt.png
  time: 6.3782s
  per call: 3.19ms

Profiling create_bundle_from_mempool()
  output written to: mempool-create-mt.png
  time: 5.9528s
  per call: 11.91ms

Profiling new_peak() (optimized)
  output written to: mempool-new-peak-mt.png
  time: 17.2157s
  per call: 8.61ms

Profiling new_peak() (reorg)
  output written to: mempool-new-peak-reorg-mt.png
  time: 16.9127s
  per call: 8.46ms
```

the profile of the most-impacted operation with sqlite backing:
![mempool-create-mt](https://user-images.githubusercontent.com/661450/221732999-e65e89ba-4956-49b7-b319-0c5a4ccbe65b.png)

## Ubuntu AMD64 (memory)

current `main`:
```
        Command being timed: "python mempool.py"
        User time (seconds): 198.14
        System time (seconds): 1.24
        Percent of CPU this job got: 109%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 3:02.68
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 78864
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 92787
        Voluntary context switches: 34294
        Involuntary context switches: 861
        Swaps: 0
        File system inputs: 0
        File system outputs: 3232
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```

with sqlite:
```
        Command being timed: "python mempool.py"
        User time (seconds): 209.45
        System time (seconds): 0.87
        Percent of CPU this job got: 111%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 3:09.48
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 79908
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 120195
        Voluntary context switches: 37254
        Involuntary context switches: 918
        Swaps: 0
        File system inputs: 0
        File system outputs: 4152
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```

## Raspberry PI (CPU)

current `main`:
```
Profiling add_spend_bundle()
  output written to: mempool-add-mt.png
  time: 18.9916s
  per call: 9.50ms

Profiling add_spend_bundle() with replace-by-fee
  output written to: mempool-replace-mt.png
  time: 18.9871s
  per call: 9.49ms

Profiling create_bundle_from_mempool()
  output written to: mempool-create-mt.png
  time: 16.9782s
  per call: 33.96ms

Profiling new_peak() (optimized)
  output written to: mempool-new-peak-mt.png
  time: 47.1240s
  per call: 23.56ms

Profiling new_peak() (reorg)
  output written to: mempool-new-peak-reorg-mt.png
  time: 45.6839s
  per call: 22.84ms
```

with sqlite:
```
Profiling add_spend_bundle()
  output written to: mempool-add-mt.png
  time: 19.0876s
  per call: 9.54ms

Profiling add_spend_bundle() with replace-by-fee
  output written to: mempool-replace-mt.png
  time: 19.0887s
  per call: 9.54ms

Profiling create_bundle_from_mempool()
  output written to: mempool-create-mt.png
  time: 21.5251s
  per call: 43.05ms

Profiling new_peak() (optimized)
  output written to: mempool-new-peak-mt.png
  time: 46.4771s
  per call: 23.24ms

Profiling new_peak() (reorg)
  output written to: mempool-new-peak-reorg-mt.png
  time: 46.4199s
  per call: 23.21ms
```