@aqk 2023-03-01
Great comment.
https://github.com/Chia-Network/chia-blockchain/pull/14657/files#diff-cad97bc75b80762f85e79ff71c5640bdeb58cfbb97992c82b08b64df958d262aR550