@aqk 2023-03-01
It would be nice to compute this from (2^63) and MAX_BLOCK_COST_CLVM * BLOCK_SIZE_LIMIT_FACTOR.

But this comment is more than sufficient.

https://github.com/Chia-Network/chia-blockchain/pull/14657/files#diff-1cd73731eeaeb06087e341a00a41d5a6e819772688cb56434f6687895da82dffR23