@wjblanke 2023-03-03
We should add some exception handling in the off case that sqlite borks, but this seems ok for now